+++ src/Ampersand.hs
@@ -3,6 +3,7 @@ module Ampersand
   , module Ampersand.Basics
   , module Ampersand.Classes
   , module Ampersand.Components
+  , module Ampersand.Daemon.Daemon
   , module Ampersand.FSpec
   , module Ampersand.Input
   , module Ampersand.Misc
@@ -14,6 +15,7 @@ import Ampersand.ADL1
 import Ampersand.Basics
 import Ampersand.Classes
 import Ampersand.Components
+import Ampersand.Daemon.Daemon
 import Ampersand.FSpec
 import Ampersand.Input
 import Ampersand.Misc